pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "halloween-leds"
version = "1.0.0"
description = "LED Light Show Controller with Music Synchronization"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["led", "wled", "music", "synchronization"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Home Automation",
]
dependencies = [
    "pygame>=2.5.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=5.13.0",
    "pytest>=7.4.0",
    "black>=23.7.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/halloween-leds"
Documentation = "https://github.com/yourusername/halloween-leds#readme"
Issues = "https://github.com/yourusername/halloween-leds/issues"

[project.scripts]
music-sync = "halloween_leds.music_sync:main"
preset-upload = "halloween_leds.wled_preset_uploader:main"

[tool.black]
line-length = 100
target-version = ["py39"]
